+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v2.2.0"></a>
+## [v2.2.0] - 2021-12-18
 ### Chore
 - removed kommentary
 
@@ -250,7 +253,8 @@
 - **indentation:** moved indentation to it's own folder
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v2.2.0...HEAD
+[v2.2.0]: https://github.com/tigorlazuardi/nvim/compare/v2.1.0...v2.2.0
 [v2.1.0]: https://github.com/tigorlazuardi/nvim/compare/v2.0.3...v2.1.0
 [v2.0.3]: https://github.com/tigorlazuardi/nvim/compare/v1.3.1...v2.0.3
 [v1.3.1]: https://github.com/tigorlazuardi/nvim/compare/v1.3.0...v1.3.1